Basic numbers logic
Numbers can be dragged and they are only accepted in the container for numbers. Appropriate log message is displayed in place of spawning a corresponding object for now. Appropriate log message is displayed if number is dragged to the wrong container.
BUG DETECTED: the first time a number is dragged to the container, nothing is being logged, as if nothing happened.