Switched everything to SO_GameObjectReference. Scale mechanics.
Classes now work exclusively with scriptable object references to game objects instead of directly with game objects.
Implemented simple scale balancing with the LeanTween library.
!! Somehow the bug where the first object that gets dropped in the scene isn't instantiated is fixed, I have no idea how it got fixed, but I'm glad it did :smile:

Next, I need to implement the removal of objects from the scale and hinge joints on the scale. Then, I think the prototype will be finished and I can move on to actual visuals.