3D object dragging logic - interfaces and implementation (Zbrojkici).
Refactored and finished the basic logic for dragging 3d objects. Currently only Zbrojkici can be dragged because numbers are text objects so I have to adjust them.
When a 3D object is picked up, a corresponding drop area is displayed. If the object is dropped within that area, it should be returned to the pool and its value deducted (TO-DO!).
Otherwise, the object is returned to its position on the scale.