Created objects for numbers. Refactored DropTarget and LevelGrid. Started with object exit behavior.
Previously numbers were just text mesh objects, which made it impossible to pick them up. Now they are actual objects with a child text mesh.
Moved most of the logic that was previously in DropTarget to LevelGrid. DropTarget kept calling LevelGrid's functions which made no sense + this way I could have other classes subscribe to only one event in only one class (LevelGrid, OnGridStateChange). The event is passing an int as an argument (1 or -1) which enables other classes to add or deduct item's value when the grid state changes.
I started with object exit behavior. Currently the logic is in LevelGrid, but it should not stay there. I created a separate class for movement paths, I still have to see how to manage this behavior.  Another TO-DO: the object currently snaps to the path starting point. The fall should be smoother.