Finished item discard.
Both item types can now be correctly removed from the scale and returned to the pool.
I added a new class - ItemObjectMover - to handle this behavior.
Animation should either be handled here or in a separate class on the same object.
Waypoints class is probably useless now, but don't remove until prototype is finished.

TO-DO: Image on drop canvas (on the scale) only becomes visible when the item is being dragged.